XBlock Integration Files/xblock-dance-discussion/requirements.txt
mysql-connector-python
sys
pkg_resources
urllib3
-e .